# Also collapse strings in patterns