@josevalim 2020-10-29
It may be that keeping those as strings would actually be the most efficient thing to do and handle them only on v3_kernel. But I believe it would require changes across `v3_core` (perhaps even syntax/parser ones).