@okeuday 2020-10-29
@josevalim Doesn't this change make matching on a prefix of a large string slower?  If the string has to be 1024 character or longer to have a slowdown, that is quite large, though I wanted to check if that is possible.