@josevalim 2020-10-29
@okeuday, this patch should not change performance/behaviour at all. The reason it doesn't affect matching is because v3_kernel expands integers with size <= 1024 back into integers of size 8 in order to build a matching tree. One way to interpret this patch is as if it was compressing AST nodes into large integers because they are cheaper to represent inside binaries.