@bjorng 2020-10-30
Looks good to me.

Have you run tests (built OTP with an updated primary bootstrap + compiler test suite)?

Have you run `scripts/diffable` to verify that the generated code didn't change? (Invoke `scripts/diffable` without arguments to print usage instructions.)