@josevalim 2020-10-30
I will do all that!

> Have you run tests (built OTP with an updated primary bootstrap + compiler test suite)?

To make sure I don't screw this up, I should:

1. Apply my changes
2. `cd lib/compiler && make`
3. `./otp_build update_primary`
4. `./otp_build all -a && ./otp_build tests`
5. Run compiler suite