@josevalim 2020-10-30
Ok, I have force pushed. It includes two fixes:

1. fixes xmerl related to utf-16 binaries
2. fixes scripts/diffable difference related to empty strings

We should be good to go now!